ntdll: Fixup the RtlWalkHeap entry pointer before subheap lookup.
The commit end of a subheap may be equal to the beginning of another
subheap, in which case find_subheap() will return that one, when we pass
it the previously returned uncommitted range lpData, and we may
effectively skip backwards in the subheap list.

This fixes a hang on starting or loading a game with Bloodrayne:
Terminal Cut.

Based on a patch from Zebediah Figura.