ntdll/tests: Update test_RtlUniform to match Windows Vista or later.
Today, test_RtlUniform() skips almost all of its tests on Windows Vista
or later.  The skipped tests only succeed in Windows XP, and fail on
Windows Vista or later.

This is because RtlUniform()'s behavior has changed, presumably due to a
bug fix shipped in Windows Vista.  It appears that RtlUniform, a linear
congruential generator, could overflow before computing the modulo in
WindoWs XP.  This is no longer the case in Windows Vista or later.

Meanwhile, we no longer support Windows XP behavior and thus do not test
for it regularly; therefore, the tests are obsolete as of 2023.

Remove obsolete tests that no longer work with any of the Windows
versions actively tested by WineHQ (as of 2023), and replace them with
updated tests that works in the Windows versions from Vista up to 10.