vbscript/tests: Test throwException with Description but *not* Source.
It occurred to me to wonder whether it was really Description, Source,
or both that results in skipping the map_vbs_exception logic.

It turns out we had it right (it's just description, and it's
null-pointer and not empty string that makes the difference).

But the fact that it's not obvious still makes it a good testcase