wintab32: Fix match_token to match XWayland devices.
XWayland tablet devices have the Wayland seat ID appended to the end of
them, preceded by a colon. This prevents match_token from properly
detecting XWayland devices and causes Wintab32 to fail to initialize.
This change allows a matched token to end in a colon, and thus allows
XWayland tablet devices to be detected correctly.

Signed-off-by: John Chadwick <john@jchw.io>