mshtml: Do not enumerate internal dynamic props.
Props allocated with dispex_get_dprop_ref or dispex_get_dynid are purely
internal to our implementation and must not be enumerated.

Note that in case of window, the props themselves become enumerable, but the
dynamic props must still be hidden, since it's the custom prop that refers
to it that must be enumerated (i.e. the DISPID must match with the custom
prop, not the underlying dynamic prop backing it, which would violate the
former DISPID obtained for the respective name).

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>