kernel32/tests: Remove tests comparing expected output in certain functions.
The Windows versions of [Read|Write]ConsoleOutput[Attribute|Character]
appear to assign an uninitialized internal DWORD to the given pointer
address before returning.

As the functions fail (intentionally) in these tests, the value of the
DWORD is never set, so the tests end up comparing the expected value of
zero with random values.

This value of this test is very limited and is best removed.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=53686