winevulkan: Generate "in_ext" definition on demand.
There are cases in StructConversionFunction.definition where we will
generate copy code for extension struct members, without emitting the
definition of "in_ext" variable used in the copy code.

This issue is triggered by mismatches in the condition that guards the
generation of the "in_ext" definitions, and the condition(s) that govern
the generation of the member copy code (e.g., in
StructConversionFunction.member_needs_copy and
VkMember.needs_conversion).

In order to avoid such mismatches and the burden of having to keep the
conditions in sync, this commit generates the definition on demand, by
checking if it's actually needed by the member copy code.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>