wined3d: Rename wined3d_context_gl_apply_fbo_state_blit() function.
To wined3d_context_gl_apply_fbo_state_explicit(). It's not really
related to blitting in principle; what it does is attaching specific
textures to the FBO instead of the d3d render targets, which was the
"original" use of FBOs in wined3d.

BTW even that original use case (currently handled by
context_state_fb()) is not using render_targets[] directly anymore and
we end up kind of abusing the blit_targets[] arrays in struct
wined3d_context_gl. Maybe we could rename that array as well.