opengl32: Avoid calling functions that have not been queried yet.
The glMapBuffer and al. functions are all duplicated between EXT/ARB and
core functions, and we reduced code duplication by redirecting the EXT
to the core functions.

The internal function table is only filled as the functions are queried,
and it causes crahes when games are using the EXT functions without
querying the core function first.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=53990