user32/tests: Don't expect WM_DISPLAYCHANGE with the default display mode.
It looks like the behavior changes when only this test is run vs when it
is run as part of the winetest suite.

This is probably because ChangeDisplaySettingsExW only sends the message
the first time it is called (since boot?).

Having other tests run before it, one of them probably changed display
settings already and we're not getting the message anymore.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=53894