user32: Fix a SetClipboardData() underflow and improve the tests.
A 0 or 1 byte Unicode string cannot be NUL terminated. That does stop
Windows 10 and older from doing so which is why the 1 byte test must be
skipped to avoid a crash in the 64-bit case.
But in such cases SetClipboardData() fails on Windows 11 and so should
Wine (instead of underflowing the buffer like Windows).
Reorganize the test data by increasing size.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=54192