wined3d: Use a string channel desc to initialize typed formats.
Avoid the need for manually set type attributes, and initialize the component
count for all such formats.

This fixes use of R11G11B10_FLOAT as a vertex attribute format with the GL
backend. Since it has no corresponding typeless format, prior to this commit the
component count was uninitialized and hence use of the format would result in a
GL error.

This fixes a GL error with Dark Souls Remastered. I did not immediately observe
any corresponding rendering errors, however.