dinput/tests: Remove the async cancel tests.
The tests are actually racy, and the async operation is cancellable in
a very short period of time after its creation, but before it started
executing.

This will be hard to test precisely and it's not related to dinput tests
in any way. Let's remove this to fix some spurious Windows failures.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=54283