kernelbase: Fix grouping repeat for number formatting.
Fixes a regression introduced by 56099a31242dcc522fb94643e70553e152c522ec.

LOCALE_SGROUPING's string and the `Grouping` field in NUMBERFMTW are
confusingly different. The former, which is what is fixed here, treats a '0'
as a repeat of the previous grouping. But a 0 at the end of the `Grouping`
field prevents it from repeating (it repeats by default otherwise) so
it's the opposite. Note that without a '0' in the LOCALE_SGROUPING string,
it shouldn't even repeat in the first place.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>