winegstreamer: Move flipping based on RGB to the frontends.
Give the backend a more simple and self-consistent API.

This commit also changes behaviour, by virtue of *not* changing some frontends.
In specific, this commit does not modify the AVI splitter, which is known to
output top-down RGB samples on Windows if the original video uses them (this is
trivial to test by modifying test.avi in quartz to use "bgra" instead of
"yuv420p"). It also does not modify the Media Foundation color converter DMO,
whose tests imply that the MF_MT_DEFAULT_STRIDE attribute is always positive.