dinput: Use a WH_CBT hook instead of WH_CALLWNDPROC.
WH_CALLWNDPROC hooks are called on every message, but we only need it
for the activation messages. WH_CBT hooks are called for a few events
only, which is much lighter, as we need to use a global hook to track
windows which belong to any other thread.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=54431