wined3d: Check for WINED3DUSAGE_LEGACY_CUBEMAP instead of checking the GL texture target.
This check was introduced in 4640be8dc834. At that point the only way for that
condition to be false was for a cube map texture.

It's not clear why cube maps are relevant here; we have no tests for this, but
it seems most expedient to just preserve the intent of the check.