dbghelp/tests: Add wrapper for EnumerateLoadedModulesW64().
Since this API sporadically fails with STATUS_INFO_LENGTH_MISMATCH
as GetLastError() (sic!) on Windows 11, retrying the call let us get
the relevant output.

No clear explanation of the cause of the failure, it's maybe generated
when modules are still loaded into child process and it detects
modification of the modules' list while enumerating all modules.

Signed-off-by: Eric Pouech <eric.pouech@gmail.com>