user32/tests: Do not modify cursor position when simulating clicks.
FVWM by default uses a focus follow mouse model so the window under the mouse cursor automatically
gets focus. Windows explorer.exe and other windows managers use click to focus model. So on FVWM,
if a test changes the cursor position, it might affects other tests that rely on a specific focus
window. Restore the cursor position after sending simulating clicks to avoid affecting other tests
unintentionally. FVWM could be configured to use a click to focus model, but right now it will make
many tests starting to succeed and other tests fail. So it seems to be too big of a change. Flaky
is added to test_SetWindowPos() because this patch makes the tests succeed on Gitlab CI but the same
tests still fails for other window managers and on TestBots.