d3d9/tests: Restore the focus window before sending SC_MAXIMIZE to it.
To match d3d8.

The original patch [1] claims

"WM_SIZE(SIZE_MAXIMIZED) is reliable here. The unreliable behavior
occured only with NOWINDOWCHANGES."

but this doesn't match my testing. Probably when that patch was submitted, this
very fix was applied to the d3d8 tests but forgotten for the d3d9 tests, and
then the wrong conclusions were drawn.

[1] https://www.winehq.org/pipermail/wine-patches/2014-December/135964.html