user32/tests: Do not stop checking the message list when skipping an optional or unsupported message.
This was either added by accident, or in an attempt to stop checking the message
list once we hit a winevent hook todo [and accidentally given too wide of a
scope]. However, the same commit also uses a global counter to ensure that only
one winevent todo is printed, so we don't need to break in the first hunk
anyway.

For trailing winevent todos, there is no global counter, so instead move the
goto inside the if block.

Fixes: a72bffe768c9fe462010d7e40e20226322fb82c7

The test marked todo here was broken by 33617af8145b6fadfe41d48968a728ff57b3064b, which was hidden due to the above commit.