winewayland.drv: Report basic monitor information.
Whenever the Wayland output display state changes (including during
initialization), update the monitor information on the Wine side to
reflect these changes.

For now all monitors are placed at 0,0 in the monitor space, and
only the current mode is reported. These deficiencies will be addressed
in upcoming commits.

We currently support a single GPU, and a single monitor for each
adapter.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>