winewayland.drv: Infer and report Windows monitor positions.
Use the xdg-output-unstable-v1 protocol to get the size of the Wayland
outputs in the compositor logical space, and use this information, along
with the logical position and size in physical pixels (i.e., the current
mode) to infer the Windows virtual screen coordinates of the monitors.

Note that there are multiple possible mappings from Wayland logical
coordinates to Windows virtual screen coordinates. We choose one
algorithm that works well for the majority of output arrangements, and
we can refine in the future as needed.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>